Rename to ControlAgent
