+++ src/Core/Command/ControlAgent.ts
@@ -1,16 +1,16 @@
 import { Maybe } from "@/Core/Language";
 import { Query } from "@/Core/Query";
 
-export interface AgentAction {
-  kind: "AgentAction";
+export interface ControlAgent {
+  kind: "ControlAgent";
   name: string;
   action: "pause" | "unpause";
 }
 
-export interface AgentActionManifest {
+export interface ControlAgentManifest {
   error: string;
   apiData: string;
   body: null;
-  command: AgentAction;
+  command: ControlAgent;
   trigger: (query: Query.SubQuery<"GetAgents">) => Promise<Maybe.Type<string>>;
 }
