+++ src/Data/Managers/ControlAgent/CommandManager.ts
@@ -1,12 +1,12 @@
 import { ApiHelper, UpdaterWithEnv } from "@/Core";
 import { CommandManagerWithEnv } from "@/Data/Common";
 
-export class AgentActionCommandManager extends CommandManagerWithEnv<"AgentAction"> {
+export class ControlAgentCommandManager extends CommandManagerWithEnv<"ControlAgent"> {
   constructor(
     private readonly apiHelper: ApiHelper,
     private readonly updater: UpdaterWithEnv<"GetAgents">
   ) {
-    super("AgentAction", ({ name, action }, environment) => {
+    super("ControlAgent", ({ name, action }, environment) => {
       return async (query) => {
         const result = await this.apiHelper.postWithoutResponse(
           `/api/v2/agent/${name}/${action}`,
