+++ src/Data/Managers/index.ts
@@ -41,4 +41,4 @@ export * from "./GetAgentProcess";
 export * from "./GetDesiredStates";
 export * from "./PromoteVersion";
 export * from "./GetVersionResources";
-export * from "./AgentAction";
+export * from "./ControlAgent";
